Move table components to table directory
