+++ ui/client/components/table/AddProcessButton.js
@@ -1,4 +1,4 @@
-import createProcessIcon from "../assets/img/create-process.svg"
+import createProcessIcon from "../../assets/img/create-process.svg"
 import React from 'react';
 
 export default class AddProcessButton extends React.Component {
