+++ ui/client/components/table/TableSelect.js
@@ -1,6 +1,6 @@
 import React from "react"
 import Select from 'react-select'
-import "../stylesheets/processes.styl"
+import "../../stylesheets/processes.styl"
 
 export default class TableSelect extends React.Component {
