+++ ui/client/containers/SubProcesses.js
@@ -12,9 +12,9 @@ import BaseProcesses from "./BaseProcesses"
 import {Glyphicon} from 'react-bootstrap'
 import ProcessUtils from "../common/ProcessUtils"
 import {nkPath} from "../config";
-import AddProcessButton from "./AddProcessButton"
-import TableSelect from "./TableSelect"
-import SearchFilter from "./SearchFilter"
+import AddProcessButton from "../components/table/AddProcessButton"
+import TableSelect from "../components/table/TableSelect"
+import SearchFilter from "../components/table/SearchFilter"
 
 class SubProcesses extends BaseProcesses {
   queries = {
